add report locking patch admin api
