+++ backend/src/v1/prisma/schema.prisma
@@ -128,6 +128,8 @@ model pay_transparency_report {
   is_unlocked                                               Boolean                            @default(true)
   reporting_year                                            Decimal                            @db.Decimal
   report_unlock_date                                        DateTime?                          @db.Timestamp(6)
+  idir_modified_username                                    String?                            @db.VarChar(255)
+  idir_modified_date                                        DateTime?                          @db.Timestamp(6)
   pay_transparency_calculated_data                          pay_transparency_calculated_data[]
   naics_code_pay_transparency_report_naics_codeTonaics_code naics_code                         @relation("pay_transparency_report_naics_codeTonaics_code", fields: [naics_code], references: [naics_code], onDelete: NoAction, onUpdate: NoAction, map: "pay_transparency_report_naics_code_fk")
   employee_count_range                                      employee_count_range               @relation(fields: [employee_count_range_id], references: [employee_count_range_id], onDelete: NoAction, onUpdate: NoAction, map: "pay_transparency_report_employee_count_range_id_fk")
