+++ backend/src/v1/services/admin-report-service.ts
@@ -7,8 +7,10 @@ import {
 } from '../types/report-search';
 import prismaReadOnlyReplica from '../prisma/prisma-client-readonly-replica';
 import { PayTransparencyUserError } from './file-upload-service';
+import prisma from '../prisma/prisma-client';
+import { LocalDateTime, ZoneId, convert } from '@js-joda/core';
 
-const reportSearchService = {
+const adminReportService = {
   /**
    * Search reports, with pagination, sorting and filtering by connecting to read replica of crunchy.
    * This gives flexibility to query with pagination and filter without any modification to this service.
@@ -68,6 +70,47 @@ const reportSearchService = {
       total: count,
       totalPages: Math.ceil(count / limit),
     };
+  },
+  /**
+   * Set report as locked/unlocked
+   * @param reportId the id of the report
+   * @param idirUsername the IDIR username 
+   * @param isUnLocked true/false to update the is_unlocked column to
+   * @returns 
+   */
+  changeReportLockStatus: async (
+    reportId: string,
+    idirUsername: string,
+    isUnLocked: boolean,
+  ) => {
+    await prisma.pay_transparency_report.findUniqueOrThrow({
+      where: { report_id: reportId },
+    });
+
+    const currentDate = convert(LocalDateTime.now(ZoneId.UTC)).toDate();
+
+    await prisma.pay_transparency_report.update({
+      where: { report_id: reportId },
+      data: {
+        is_unlocked: isUnLocked,
+        report_unlock_date: currentDate,
+        idir_modified_date: currentDate,
+        idir_modified_username: idirUsername,
+      },
+    });
+
+    return prisma.pay_transparency_report.findUnique({
+      where: { report_id: reportId },
+      include: {
+        employee_count_range: true,
+        pay_transparency_company: {
+          select: {
+            company_id: true,
+            company_name: true,
+          },
+        },
+      },
+    });
   },
   /**
    * Convert JSON object to Prisma specific query object
@@ -132,4 +175,4 @@ const convertSortToPrismaFormat = (
   });
 };
 
-export { reportSearchService };
+export { adminReportService };
